Conflicts:
	2012/Gpp/Gpp/MovableObject.cs

+++ 2012/Gpp/Gpp/GameObject.cs
@@ -12,7 +12,7 @@ namespace Gpp
         /// <summary>
         /// (m*m*m)/kg/s/s
         /// </summary>
-        public const double GravitationConstant = 6.673E-11;
+        public const float GravitationConstant = 6.673E-11f;
 
         public SupermassiveGame Game { get; private set; }
 
@@ -26,7 +26,7 @@ namespace Gpp
         /// <summary>
         /// KG
         /// </summary>
-        public double Mass { get; protected set; }
+        public float Mass { get; protected set; }
 
         public GameObject(SupermassiveGame game)
         {
